Allow for pattern detect to call or not call N numbers
